Non-macro form of LENGTH() reinstated (for use by packages).
git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1473 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/include/Rinternals.h
@@ -216,7 +216,7 @@ void (SET_S4_OBJECT)(SEXP x);
 void (UNSET_S4_OBJECT)(SEXP x);
 
 /* Vector Access Functions */
-size_t  (LENGTH)(SEXP x);
+int  (LENGTH)(SEXP x);
 int  (TRUELENGTH)(SEXP x);
 void (SETLENGTH)(SEXP x, int v);
 void (SET_TRUELENGTH)(SEXP x, int v);
